+++ CHANGELOG.md
@@ -49,6 +49,27 @@
 
 ### Added
 
+* Support for WN-LMF 1.4 ([#260])
+  - Sense ordering: `index` on `<LexicalEntry>` and `n` on `<Sense>`
+  - New sense relations:
+    - `metaphor`
+    - `has_metaphor`
+    - `metonym`
+    - `has_metonym`
+    - `agent`
+    - `material`
+    - `event`
+    - `instrument`
+    - `location`
+    - `by_means_of`
+    - `undergoer`
+    - `property`
+    - `result`
+    - `state`
+    - `uses`
+    - `destination`
+    - `body_part`
+    - `vehicle`
 * `wn.compat` namespace (see [#55])
 * `wn.compat.sensekey` module ([#55]) with methods:
   - `sense_key_getter()`
@@ -67,6 +88,7 @@
   calculations ([#255])
 * `wn.add()` no longer requires `partOfSpeech` on synsets; this was
   not a requirement of WN-LMF nor was it enforced in the database
+* `wn.export()` defaults to `version="1.4"` instead of `"1.0"`
 
 
 ## [v0.12.0]
@@ -716,6 +738,7 @@ the https://github.com/nltk/wordnet/ code which had been effectively
 abandoned, but this is an entirely new codebase.
 
 
+[v0.14.0]: ../../releases/tag/v0.14.0
 [v0.13.0]: ../../releases/tag/v0.13.0
 [v0.12.0]: ../../releases/tag/v0.12.0
 [v0.11.0]: ../../releases/tag/v0.11.0